Split function related code out of internal_rep.rs into its own file
